+++ functions/src/index.ts
@@ -3,4 +3,5 @@ export {leaveHousehold, leaveHouseholdHttp} from "./leaveHousehold";
 export {sendHouseholdInvitation, sendHouseholdInvitationHttp} from "./sendHouseholdInvitation";
 export {createSubscription} from "./stripe";
 export {migrateHouseholdClaims, migrateHouseholdClaimsHttp} from "./migrateHouseholdClaims";
+export {getNutritionData} from "./nutrition";
 // export {createPayPalSubscription, approvePayPalSubscription} from "./paypal"; // Temporarily disabled